refactor: remove metrics.cli and qa_generation.cli, move create_retrieval_function to metrics.runner
